// src/module/game/application/games/true-or-false/enums/TrueOrFalse.enums.ts
export enum TrueOrFalseInputNames {
  rollDice = 'rollDice',
  selectTrueOrFalseOption = 'selectTrueOrFalseOption',
  endTurn = 'endTurn',
}
export enum TrueOrFalseStates {
  playerDiceRoll = 'playerDiceRoll',
  playerStory = 'playerStory',
  showingScores = 'showingScores',
  waitingForEndTurn = 'waitingForEndTurn',
}

export enum TrueOrFalseStory {
  true = 'true',
  false = 'false',
  empty = '',
}

export enum TrueOrFalseEvents {
  winnerGame = 'winnerGame',
}